requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.3
Pillow==11.0.0
reportlab==4.0.7
python-dotenv==1.0.0
PyQt5==5.15.10
matplotlib==3.9.0
